requirements.txt
streamlit
ebird-api
beautifulsoup4
pandas
requests
python-dotenv
streamlit_javascript